+++ src/isObject.ts
@@ -1,3 +1,3 @@
-export default function isObject(value: any): value is object {
+export default function isObject(value: unknown): value is object {
   return !!value && typeof value === 'object' && !Array.isArray(value);
 }