+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated to `@babel/runtime-corejs3`, which uses Babel v7.4 and CoreJS v3.
 - TS: Updated `Blueprint` to remove optional properties (acts like `Required`).
+- TS: Replaced some usage of `any` with `unknown`.
 
 # 2.1.1 - 2018-02-07